# superops/docs/swagger.yaml
definitions:
...
  modules.Kubecmds:
    properties:
      comname:
        type: string
      compath:
        type: string
      comvers:
        type: string
      ctime:
        type: string
      id:
        type: integer
      state:
        type: integer
    type: object
  modules.KubernetesConfigs:
    properties:
      comname:
        type: string
      confpath:
        type: string
      ctime:
        type: string
      id:
        type: integer
      kconfig:
        type: string
      kcontexts:
        type: string
      kenvrole:
        type: integer
      kname:
        type: string
      state:
        type: integer
    type: object
...
  modules.Projects:
    properties:
      State:
        type: integer
      context:
        type: string
      ctime:
        type: string
      cuser:
        type: string
      id:
        type: integer
      jenkins:
        type: string
      project:
        type: string
      repository:
        type: string
    type: object
...
  v1.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
  v1.UserlistSelectType:
    properties:
      lines:
        type: integer
      offset:
        type: integer
      orderby:
        type: string
      stype:
        type: integer
    type: object
host: 127.0.0.1:8080
info:
  contact:
    name: API Support
    url: http://127.0.0.1:8080/api/v1/swagger/index.html
  description: Superops project api document
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Superops-system
  version: "1.0"
paths:
  /api/v1/Project/controlk8sproject:
    get:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/Project/controlk8sproject?project=xxx&do=1|2
        project：项目名称
        do：1表示激活项目，2表示禁用项目
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Projects名称
        in: query
        name: project
        required: true
        type: string
      - description: 操作
        in: query
        name: do
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 激活、禁用k8s项目
      tags:
      - Project
  /api/v1/Project/createk8sproject:
    post:
      consumes:
      - application/json
      description: |-
        格式：POST /api/v1/Project/createk8sproject
        不可为空字段：project、jenkins、repository
        jenkins获取接口：/api/v1/service/getjenkinsnamelist
        repository获取接口：/api/v1/service/getharbornamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Projects参数
        in: body
        name: modules.Projects
        required: true
        schema:
          $ref: '#/definitions/modules.Projects'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 新增k8s项目
      tags:
      - Project
  /api/v1/Project/getk8sprojectlist:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/Project/getk8sprojectlist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询k8s项目列表(管理员表格)（此接口会刷新项目列表缓存）
      tags:
      - Project
  /api/v1/Project/getk8sprojectnamelist:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/Project/getk8sprojectnamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 管理员查询k8s项目名称列表（此接口会刷新项目列表缓存）
      tags:
      - Project
...
  /api/v1/Project/selectk8sprojectinfo:
    get:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/Project/selectk8sprojectinfo?project=xxx
        project：项目名称
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Projects名称
        in: query
        name: project
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询k8s项目信息
      tags:
      - Project
  /api/v1/Project/updatek8sproject:
    post:
      consumes:
      - application/json
      description: |-
        格式：POST /api/v1/Project/updatek8sproject
        不可为空字段：project
        jenkins获取接口：/api/v1/service/getjenkinsnamelist
        repository获取接口：/api/v1/service/getharbornamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Projects参数
        in: body
        name: modules.Projects
        required: true
        schema:
          $ref: '#/definitions/modules.Projects'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 更新k8s项目
      tags:
      - Project
  /api/v1/Project/usergetk8sprojectnamelist:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/Project/usergetk8sprojectnamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 用户查询k8s项目名称列表
      tags:
      - Project
...
  /api/v1/cmdb/controlkubeconfig:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/cmdb/controlkubeconfig?kubeconfig=xxx&do=enable|disable
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: kubeconfig名称
        in: query
        name: kubeconfig
        required: true
        type: string
      - description: 使用enable或disable来启停
        in: query
        name: do
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 启用或禁用kubeconfig文件(未完待续)
      tags:
      - cmdb
  /api/v1/cmdb/createkubeconfig:
    post:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/cmdb/createkubeconfig
        不可为空字段：kname、kenvrole、kconfig、comname
        hcomname获取接口：/api/v1/commands/getcommandsnamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: kubeconfig参数
        in: body
        name: modules.KubernetesConfigs
        required: true
        schema:
          $ref: '#/definitions/modules.KubernetesConfigs'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 添加kubeconfig文件
      tags:
      - cmdb
  /api/v1/cmdb/deletekubeconfig:
    get:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/cmdb/deletekubeconfig?kubeconfig=xxx
        kubeconfig: 配置名称
        获取kubeconfig列表接口：/api/v1/cmdb/getkubeconfiglist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: kubeconfig名称
        in: query
        name: kubeconfig
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 删除kubeconfig文件
      tags:
      - cmdb
  /api/v1/cmdb/getkubeconfig:
    get:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/cmdb/getkubeconfig?kubeconfig=xxx
        kubeconfig: 配置名称
        名称列表获取接口：GET /api/v1/cmdb/getkubeconfiglist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: kubeconfig名称
        in: query
        name: kubeconfig
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取kubeconfig文件详情(单条详情)
      tags:
      - cmdb
  /api/v1/cmdb/getkubeconfiglist:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/cmdb/getkubeconfiglist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取kubeconfig表格(用于管理页面表格展示)
      tags:
      - cmdb
...
  /api/v1/cmdb/updatekubeconfig:
    post:
      consumes:
      - application/json
      description: 格式：GET /api/v1/cmdb/updatekubeconfig
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: kubeconfig配置信息
        in: body
        name: modules.KubernetesConfigs
        required: true
        schema:
          $ref: '#/definitions/modules.KubernetesConfigs'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 修改kubeconfig文件
      tags:
      - cmdb
  /api/v1/commands/controlcommand:
    get:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/commands/controlcommand?command=xxx&type=0|1|2|3
        command: 命令名称
        type: 0、查询无命令文件的命令，1、获取激活中的命令，2、获取禁用的命令，3、获取所有命令
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: 查询类型。0、查询无命令文件的命令，1、获取激活中的命令，2、获取禁用的命令，3、获取所有命令
        in: query
        name: type
        required: true
        type: string
      - description: 需要禁用的命令名称
        in: query
        name: command
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 命令启用或禁用
      tags:
      - Commands
  /api/v1/commands/createcommand:
    post:
      consumes:
      - application/json
      description: 格式：GET /api/v1/commands/createcommand
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: 命令信息
        in: body
        name: modules.Kubecmds
        required: true
        schema:
          $ref: '#/definitions/modules.Kubecmds'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 新增命令
      tags:
      - Commands
  /api/v1/commands/createfile:
    post:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/commands/createfile?comname=xxx file=文件
        注意：需要先成功创建命令后，才能上传命令执行文件(命令必须是激活状态才能上传可执行文件)
        comname是命令名称，获取接口：/api/v1/commands/getcommandsnamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: 命令名称
        in: query
        name: comname
        required: true
        type: string
      - description: 命令文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 上传命令执行文件
      tags:
      - Commands
  /api/v1/commands/delcommands:
    get:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/commands/delcommands?command=xxx
        command: 命令名称
        command获取接口：/api/v1/commands/getcommandsnamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: 删除命令名称
        in: query
        name: command
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 删除命令
      tags:
      - Commands
  /api/v1/commands/getcommandslist:
    get:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/commands/getcommandslist?type=0|1|2|3
        type: 0、查询无命令文件的命令，1、获取激活中的命令，2、获取禁用的命令，3、获取所有命令
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: 查询类型
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取命令列表(详情)
      tags:
      - Commands
  /api/v1/commands/getcommandsnamelist:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/commands/getcommandsnamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取激活命令名称列表
      tags:
      - Commands
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: |-
        可以使用格式:
        用户名+密码
        手机号+密码
        邮箱+密码
        例如：{"username":"","password":""}
      parameters:
      - description: username、password用于登录验证
        in: body
        name: modules.Users
        required: true
        schema:
          $ref: '#/definitions/modules.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 用户登录接口
      tags:
      - Users
  /api/v1/register:
    post:
      description: 必须参数：用户名、密码、手机、邮箱
      parameters:
      - description: 用户注册
        in: body
        name: modules.Users
        required: true
        schema:
          $ref: '#/definitions/modules.Users'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 用户注册接口
      tags:
      - Users
  /api/v1/service/Updateharbor:
    post:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/service/Updateharbor
        必须字段：Harbor(harbor名称)
        可选字段：Hurl(harbor地址)、Huser(harbor用户)、Hpass(harbor密码)
        可选字段：Dimgtimeout(开发镜像保留的时间默认7天，0天表示不过期需要手动清理harbor，单位是天)
        可选字段：Timgtimeout(测试镜像保留的时间默认30天，0天表示不过期需要手动清理harbor，单位是天)
        可选字段：Timgtimeout(生产镜像保留的时间默认90天，0天表示不过期需要手动清理harbor，单位是天)
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Harbors参数
        in: body
        name: modules.Harbors
        required: true
        schema:
          $ref: '#/definitions/modules.Harbors'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 修改harbor
      tags:
      - service
...
  /api/v1/service/createharbor:
    post:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/service/createharbor
        必须字段：Harbor(harbor名称)、Hurl(harbor地址)、Huser(harbor用户)、Hpass(harbor密码)
        可选字段：Dimgtimeout(开发镜像保留的时间默认7天，0天表示用户过期需要手动清理harbor，单位是天)
        可选字段：Timgtimeout(测试镜像保留的时间默认30天，0天表示用户过期需要手动清理harbor，单位是天)
        可选字段：Timgtimeout(生产镜像保留的时间默认90天，0天表示用户过期需要手动清理harbor，单位是天)
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Harbors参数
        in: body
        name: modules.Harbors
        required: true
        schema:
          $ref: '#/definitions/modules.Harbors'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 添加harbor
      tags:
      - service
  /api/v1/service/createjenkins:
    post:
      consumes:
      - application/json
      description: 格式：GET /api/v1/service/createjenkins
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: jenkins信息
        in: body
        name: modules.Jenkins
        required: true
        schema:
          $ref: '#/definitions/modules.Jenkins'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 添加jenkins
      tags:
      - service
  /api/v1/service/deleteharbor:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/service/deleteharbor?harbor=xxx
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Harbors名称
        in: query
        name: harbor
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 删除harbor(未完成，删除前需要查询所有项目都没有使用此harbor)
      tags:
      - service
  /api/v1/service/getharborinfo:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/service/getharborinfo?harbor=xxx
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Harbors名称
        in: query
        name: harbor
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询harbor单条详情
      tags:
      - service
  /api/v1/service/getharborlist:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/service/getharborlist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Harbors名称
        in: query
        name: harbor
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询harbor列表(管理员页面展示表格)
      tags:
      - service
  /api/v1/service/getharbornamelist:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/service/getharbornamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: Harbors名称
        in: query
        name: harbor
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询harbor名称列表
      tags:
      - service
...
  /api/v1/service/getjenkinsjobslist:
    get:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/service/getjenkinsjobslist?jenkins=xxx
        jenkins:jenkins名称
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: jenkins名称
        in: query
        name: jenkins
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询jenkins job列表
      tags:
      - service
  /api/v1/service/getjenkinslist:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/service/getjenkinslist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询jenkins列表
      tags:
      - service
  /api/v1/service/getjenkinsnamelist:
    get:
      consumes:
      - application/json
      description: 格式：GET /api/v1/service/getjenkinsnamelist
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询jenkins名称列表
      tags:
      - service
  /api/v1/service/jenkinsinfo:
    get:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/service/jenkinsinfo?jenkins=xxx
        jenkins：查询jenkins的名称
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: 查询的jenkins名称
        in: query
        name: jenkins
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 查询单条jenkins详情
      tags:
      - service
...
  /api/v1/service/updatejenkins:
    post:
      consumes:
      - application/json
      description: |-
        格式：GET /api/v1/service/controljenkins?jenkins=xxx&type=1|2|3&to
        必须字段：
        jenkins：指定需要修改的jenkins名称
        type：1表示启用，2表示禁用，3表示迁移(to=xxx必须，用于指定迁移的目标jenkins)
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: jenkins名称
        in: query
        name: jenkins
        required: true
        type: string
      - description: 操作
        in: query
        name: type
        required: true
        type: string
      - description: 目标jenkins
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: jenkins状态控制(未完待续)
      tags:
      - service
...
  /api/v1/users/admingetlist:
    post:
      consumes:
      - application/json
      description: |-
        请求格式：{"stype":0,"lines":20,"offset":1,"orderby":"id"}
        stype为查询类型：0、查询新注册用户,1、查询激活用户，2、查询禁用用户，3、查询所有用户
        lines为查询条数，默认20条，最多不能查询100条
        offset查询页码，默认查询第1页
        orderby查询排序列，默认为id
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: 查询用户列表参数
        in: body
        name: UserlistSelectType
        required: true
        schema:
          $ref: '#/definitions/v1.UserlistSelectType'
      produces:
      - application/json
      responses:
        "200":
          description: ok" "返回用户信息
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取用户列表
      tags:
      - UsersAdmin
  /api/v1/users/chageinfo:
    post:
      consumes:
      - application/json
      description: 修改用户信息，可修改：签名、手机号、邮箱、描述信息
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: 修改用户信息
        in: body
        name: modules.UsersInfo
        required: true
        schema:
          $ref: '#/definitions/modules.UsersInfo'
      produces:
      - application/json
      responses:
        "200":
          description: ok" "返回用户信息
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 用户信息修改
      tags:
      - Users
  /api/v1/users/chagespass:
    post:
      consumes:
      - application/json
      description: |-
        修改用户密码，直接在url参数中提交密码即可，例如：password=xxxxx
        用户修改密码后需要重新登录
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: password参数
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 用户修改密码
      tags:
      - Users
  /api/v1/users/controls:
    get:
      consumes:
      - application/json
      description: |-
        启用和禁用单个用户
        格式：GET /api/v1/users/controls?user=xxx&do=1
        user：输入用户名
        do：操作，1表示启用，2表示禁用
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - description: 启用/禁用用户名
        in: query
        name: user
        required: true
        type: string
      - description: 启用：1，禁用：2
        in: query
        name: do
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok" "返回用户信息
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 单用户启用/禁用
      tags:
      - UsersAdmin
    post:
      consumes:
      - application/json
      description: |-
        用户批量启用和禁用
        格式：POST /api/v1/users/controls {"doenable":["x","x","x"],"dodisable":["x","x","x"]}
        doenable：激活用户列表
        dodisable：禁用用户列表
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      - in: query
        items:
          type: string
        name: dodisable
        type: array
      - in: query
        items:
          type: string
        name: doenable
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 批量用户启用/禁用
      tags:
      - UsersAdmin
  /api/v1/users/createapitoken:
    get:
      description: 请求格式：GET /api/v1/users/createapitoken
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 用户创建API Token
      tags:
      - Users
  /api/v1/users/deleteapitoken:
    get:
      description: 请求格式：GET /api/v1/users/deleteapitoken
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 用户删除API Token
      tags:
      - Users
  /api/v1/users/disableapitoken:
    get:
      description: 请求格式：GET /api/v1/users/disableapitoken
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 用户停止API Token
      tags:
      - Users
  /api/v1/users/enableapitoken:
    get:
      description: 请求格式：GET /api/v1/users/enableapitoken
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 用户开启API Token
      tags:
      - Users
...
  /api/v1/users/getlists:
    get:
      description: 获取用户列表
      parameters:
      - description: 用户token
        in: header
        name: Access-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/v1.Response'
      summary: 获取用户列表
      tags:
      - Users
...
swagger: "2.0"